requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
SQLAlchemy==2.0.36
Werkzeug==3.1.3
pytest==8.3.3
python-dotenv==1.0.1
responses==0.25.3
Flask-Compress==1.18